# services/scoring/openapi.yaml
openapi: 3.1.0
info:
  title: OCR Kartu Keluarga Scoring API
  description: 'Pipeline step 4 (last) for Indonesian Kartu Keluarga: P(correct) per contract field, after
    fusing the two structuring scores. There is no document score and no approve/reject decision -- the
    threshold belongs to the caller. **Async pipeline:** the structuring service POSTs /v1/scoring/jobs
    and gets 202; this service scores in the background, stores the result, and POSTs the stage callback
    carrying the final result to the orchestrator. /v1/scoring/confidence is the same work, synchronous.
    All endpoints except /health require an X-API-Key header.


    ## Conventions (the same for every nlm-k2 service)


    **Authentication.** Every endpoint except `/health`, `/ready` and `/metrics` requires the header `X-API-Key`.

    A missing or wrong key answers `401`. The services are reachable only from inside the cluster.


    **Envelope.** Every JSON response, success or error, has the same shape:

    `{status_code, status_desc, message, data, errors, request_id}`. `status_code` always equals the HTTP

    status. On success `errors` is null; on error `data` is null.


    **Errors.** `errors` is machine-readable: `VALIDATION_ERROR` for `422`, otherwise the same text as

    `message`. `400` = the request or the document is unusable (do not retry unchanged). `401` = API key.

    `404` = unknown request_id. `409` = request_id already processed (legacy contract only). `422` = a

...

...


...

...

...


...

...

...
servers:
...
paths:
...
  /v1/scoring/jobs:
    post:
      tags:
      - Pipeline
      summary: Hand a structured document to the scoring stage (last)
      description: '**Step 4 of the pipeline, asynchronous, the last stage. Called by the structuring
        service, not by the orchestrator.**


        Records the job (`scoring_jobs`, idempotent per request_id), answers **202 immediately**, then
        in the background: builds the ML team''s scoring payload from the chained guardrails + OCR + structuring
        results, runs the trust model, stores the result (`scoring_results`), and POSTs the `SCORING`
        callback, which carries the **final result** of the request.


        The outcome is two per-field confidences. There is no document-level score and no approve / reject
        decision: thresholds belong to the orchestrator.'
      operationId: submitScoringJob
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScoringJobRequest'
        required: true
      responses:
        '202':
          description: The job was accepted (or already existed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobAcceptedResponse'
              examples:
                accepted:
                  summary: New job
                  value:
                    status_code: 202
                    status_desc: Accepted
                    message: Accepted
                    data:
                      request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
                      stage: SCORING
                      status: PROCESSING
                      duplicate: false
                    request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
                duplicate:
                  summary: 'Same request_id sent again: nothing is re-run'
                  value:
                    status_code: 202
                    status_desc: Accepted
                    message: Accepted
                    data:
                      request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
                      stage: SCORING
                      status: DONE
                      duplicate: true
                    request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
        '401':
          description: Missing or invalid X-API-Key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status_code: 401
                status_desc: Unauthorized
                message: Invalid or missing API key
                errors: Invalid or missing API key
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status_code: 422
                status_desc: Unprocessable Entity
                message: 'body.request_id: Field required'
                errors: VALIDATION_ERROR
      security:
      - APIKeyHeader: []
  /v1/scoring/jobs/{request_id}:
    get:
      tags:
      - Pipeline
      summary: Status and result of the scoring stage
      description: 'Status of this stage only and, once `DONE`, the two confidences plus the exact payload
        that was scored. Internal: the orchestrator reads it (while it waits, and for its `GET /v1/extract-ocr/{request_id}`);
        the central orchestrator receives the final result in the callback. Also useful to audit a score.'
      operationId: getScoringJob
      security:
      - APIKeyHeader: []
      parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: string
          title: Request Id
      responses:
        '200':
          description: The job exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScoringJobStatusResponse'
              examples:
                done:
                  summary: Finished
                  value:
                    status_code: 200
                    status_desc: OK
                    message: Success
                    data:
                      request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
                      stage: SCORING
                      created_at: '2026-09-18T04:00:01+00:00'
                      status: DONE
                      result:
                        nomor_kk: 0.9806
                        name_confidence: 0.9948
                        payload:
                          structuring: '{...the §7.3 structuring result, whole...}'
                          guardrail_probability: 0.0287
                          guardrail_verdict: accepted
                          avg_doc_score: 0.814
                          min_doc_score: 0.2822
                          text_regions_count: 177
                      updated_at: '2026-09-18T04:00:01+00:00'
                    request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
                failed:
                  summary: Failed
                  value:
                    status_code: 200
                    status_desc: OK
                    message: Success
                    data:
                      request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
                      stage: SCORING
                      created_at: '2026-09-18T04:00:01+00:00'
                      status: FAILED
                      error_message: 'Unsupported document_type: ktp. Supported: [''kk'']'
                      updated_at: '2026-09-18T04:00:01+00:00'
                    request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
        '401':
          description: Missing or invalid X-API-Key
          content:
            application/json:
              example:
                status_code: 401
                status_desc: Unauthorized
                message: Invalid or missing API key
                errors: Invalid or missing API key
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No scoring job for this request_id
          content:
            application/json:
              example:
                status_code: 404
                status_desc: Not Found
                message: 'No SCORING job found for request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0'
                errors: 'No SCORING job found for request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0'
                request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              example:
                status_code: 422
                status_desc: Unprocessable Entity
                message: 'path.request_id: Field required'
                errors: VALIDATION_ERROR
                request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/scoring/outbox:
    get:
      tags:
      - Pipeline
      summary: Backlog of this stage's undelivered callbacks and hand-offs
      description: 'Operational view of the transactional outbox (`PIPELINE_OUTBOX`): how many callbacks
        and hand-offs of this stage are still to be delivered, how old the oldest one is, and how many
        were given up on (dead letters). No other service watches the `pipeline_outbox` table, so this
        is where a monitor should look; the relay also logs a warning when the backlog is older than `PIPELINE_OUTBOX_STALE_AFTER_SECONDS`
        or a dead letter is waiting.'
      operationId: getScoringOutboxStatus
      responses:
        '200':
          description: The backlog of this stage
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutboxStatusResponse'
              examples:
                idle:
                  summary: Nothing waiting
                  value:
                    status_code: 200
                    status_desc: OK
                    message: Success
                    data:
                      enabled: true
                      stage: SCORING
                      pending: 0
                      retrying: 0
                      dead_letters: 0
                    request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
                backlog:
                  summary: The orchestrator has been down for a while
                  value:
                    status_code: 200
                    status_desc: OK
                    message: Success
                    data:
                      enabled: true
                      stage: SCORING
                      pending: 12
                      retrying: 12
                      oldest_pending_seconds: 754.3
                      dead_letters: 1
                    request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
                disabled:
                  summary: '`PIPELINE_OUTBOX` is off'
                  value:
                    status_code: 200
                    status_desc: OK
                    message: Success
                    data:
                      enabled: false
                      stage: SCORING
                      pending: 0
                      retrying: 0
                      dead_letters: 0
                    request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
        '401':
          description: Missing or invalid X-API-Key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status_code: 401
                status_desc: Unauthorized
                message: Invalid or missing API key
                errors: Invalid or missing API key
      security:
      - APIKeyHeader: []
  /v1/scoring/outbox/release:
    post:
      tags:
      - Pipeline
      summary: Send this stage's dead letters again
      description: 'Puts the dead letters of this stage back in the outbox queue, due immediately, so
        the relay delivers them again: use it after the receiver was fixed (for example the orchestrator''s
        callback contract, or a next stage that was answering 4xx). Optional query `request_id` limits
        it to one request. A message that fails again goes through the same retry and dead-letter rules.
        `409` when `PIPELINE_OUTBOX` is off on this service.'
      operationId: releaseScoringOutbox
      security:
      - APIKeyHeader: []
      parameters:
      - name: request_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Request Id
      responses:
        '200':
          description: Dead letters released
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutboxReleaseResponse'
              examples:
                released:
                  summary: One request
                  value:
                    status_code: 200
                    status_desc: OK
                    message: Success
                    data:
                      stage: SCORING
                      request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
                      released: 1
                    request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
                nothing:
                  summary: Nothing to release
                  value:
                    status_code: 200
                    status_desc: OK
                    message: Success
                    data:
                      stage: SCORING
                      released: 0
                    request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
        '401':
          description: Missing or invalid X-API-Key
          content:
            application/json:
              example:
                status_code: 401
                status_desc: Unauthorized
                message: Invalid or missing API key
                errors: Invalid or missing API key
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: The outbox is off on this service
          content:
            application/json:
              example:
                status_code: 409
                status_desc: Conflict
                message: 'PIPELINE_OUTBOX is off: there is no outbox to release'
                errors: 'PIPELINE_OUTBOX is off: there is no outbox to release'
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              example:
                status_code: 422
                status_desc: Unprocessable Entity
                message: 'query.request_id: Input should be a valid string'
                errors: VALIDATION_ERROR
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorResponse:
      properties:
        status_code:
          type: integer
          title: Status Code
          description: Same as the HTTP status code
          examples:
          - 400
        status_desc:
          type: string
          title: Status Desc
          description: Reason phrase of `status_code`
          examples:
          - Bad Request
        message:
          type: string
          title: Message
          description: Human-readable cause; safe to log
          examples:
          - Uploaded file is empty
        data:
          type: 'null'
          title: Data
          description: Always null on error
        errors:
          type: string
          title: Errors
          description: Machine-readable cause. `VALIDATION_ERROR` for 422 (branch on this, the message
            names a different field every time); for every other status it repeats `message`
          examples:
          - Uploaded file is empty
        request_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Request Id
          description: Present when the error is tied to a known request_id
          examples:
          - REQ_9cb01af2-493d-446d-b191-af120333f6d0
      type: object
      required:
      - status_code
      - status_desc
      - message
      - errors
      title: ErrorResponse
      description: Every error response.
...
    GuardrailsResult:
      properties:
        passed:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Passed
          description: 'true: the document may proceed to OCR'
          examples:
          - true
        reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Reason
          description: Why it was rejected, in Indonesian; null when passed
          examples:
          - null
        document:
          anyOf:
          - $ref: '#/components/schemas/GuardrailsDocument'
          - type: 'null'
          description: The verdict and its numbers
      additionalProperties: true
      type: object
      title: GuardrailsResult
      description: The guardrails report (`data` of the guardrails service's POST /v1/guardrails/check,
        sent on by the orchestrator), forwarded unchanged down the pipeline.
...
    JobAcceptedResponse:
      properties:
        status_code:
          type: integer
          title: Status Code
          description: Same as the HTTP status code
          default: 202
          examples:
          - 202
        status_desc:
          type: string
          title: Status Desc
          description: Reason phrase of `status_code`
          default: Accepted
          examples:
          - Accepted
        message:
          type: string
          title: Message
          description: Human-readable outcome
          default: Accepted
          examples:
          - Accepted
        errors:
          type: 'null'
          title: Errors
          description: Always null on success
        request_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Request Id
          description: 'The request_id this response belongs to: the one sent in the form / body / path,
            otherwise the `X-Request-ID` header, otherwise one generated by the service'
          examples:
          - REQ_9cb01af2-493d-446d-b191-af120333f6d0
        data:
          $ref: '#/components/schemas/JobAccepted'
      type: object
      required:
      - data
      title: JobAcceptedResponse
      description: Envelope of `POST /v1/<stage>/jobs`.
...
    OcrPayload:
      properties:
        engine:
          anyOf:
          - type: string
          - type: 'null'
          title: Engine
          description: OCR backend that produced the boxes
          examples:
          - kk_ocr
        model:
          anyOf:
          - type: string
          - type: 'null'
          title: Model
          description: Model identity reported by the OCR backend
          examples:
          - PP-OCRv5_server_det+PP-OCRv5_server_rec
        elapsed_ms:
          anyOf:
          - type: number
          - type: 'null'
          title: Elapsed Ms
          description: Time spent in the OCR model
          examples:
          - 4182.7
        text_regions_count:
          type: integer
          minimum: 0.0
          title: Text Regions Count
          description: Number of text boxes; may be 0
          default: 0
          examples:
          - 177
        avg_doc_score:
          anyOf:
          - type: number
            maximum: 1.0
            minimum: 0.0
          - type: 'null'
          title: Avg Doc Score
          description: Mean of `texts[].score`; null when `texts` is empty rather than 0
          examples:
          - 0.814
        min_doc_score:
          anyOf:
          - type: number
            maximum: 1.0
            minimum: 0.0
          - type: 'null'
          title: Min Doc Score
          description: Lowest of `texts[].score`; null when `texts` is empty
          examples:
          - 0.2822
        texts:
          items:
            $ref: '#/components/schemas/OcrBoxPayload'
          type: array
          title: Texts
          description: 'Text boxes. May be EMPTY: an image with no readable text must reach the structuring
            rules to be rejected there, so neither this model nor the structuring job endpoint may require
            at least one'
      additionalProperties: true
      type: object
      title: OcrPayload
      description: Result of the OCR stage (`ocr_results`), forwarded to structuring and scoring.
...
    OutboxReleaseResponse:
      properties:
        status_code:
          type: integer
          title: Status Code
          description: Same as the HTTP status code
          default: 200
          examples:
          - 200
        status_desc:
          type: string
          title: Status Desc
          description: Reason phrase of `status_code`
          default: OK
          examples:
          - OK
        message:
          type: string
          title: Message
          description: Human-readable outcome
          default: Success
          examples:
          - Success
        errors:
          type: 'null'
          title: Errors
          description: Always null on success
        request_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Request Id
          description: 'The request_id this response belongs to: the one sent in the form / body / path,
            otherwise the `X-Request-ID` header, otherwise one generated by the service'
          examples:
          - REQ_9cb01af2-493d-446d-b191-af120333f6d0
        data:
          $ref: '#/components/schemas/OutboxRelease'
      type: object
      required:
      - data
      title: OutboxReleaseResponse
      description: Envelope of `POST /v1/<stage>/outbox/release`.
...
    OutboxStatusResponse:
      properties:
        status_code:
          type: integer
          title: Status Code
          description: Same as the HTTP status code
          default: 200
          examples:
          - 200
        status_desc:
          type: string
          title: Status Desc
          description: Reason phrase of `status_code`
          default: OK
          examples:
          - OK
        message:
          type: string
          title: Message
          description: Human-readable outcome
          default: Success
          examples:
          - Success
        errors:
          type: 'null'
          title: Errors
          description: Always null on success
        request_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Request Id
          description: 'The request_id this response belongs to: the one sent in the form / body / path,
            otherwise the `X-Request-ID` header, otherwise one generated by the service'
          examples:
          - REQ_9cb01af2-493d-446d-b191-af120333f6d0
        data:
          $ref: '#/components/schemas/OutboxStatus'
      type: object
      required:
      - data
      title: OutboxStatusResponse
      description: Envelope of `GET /v1/<stage>/outbox`.
...
    ScoringJobRequest:
      properties:
        request_id:
          type: string
          minLength: 1
          title: Request Id
          description: request_id of the pipeline run
          examples:
          - REQ_9cb01af2-493d-446d-b191-af120333f6d0
        document_type:
          type: string
          title: Document Type
          description: Only `kk` is supported; anything else fails the job
          default: kk
          examples:
          - kk
        guardrails:
          anyOf:
          - $ref: '#/components/schemas/GuardrailsResult'
          - type: 'null'
          description: Guardrails result submitted with the OCR job; returned unchanged in the final result
        ocr:
          anyOf:
          - $ref: '#/components/schemas/OcrPayload'
          - type: 'null'
          description: Result of the OCR stage; its aggregates feed the model as document-level features
        structuring:
          anyOf:
          - $ref: '#/components/schemas/StructuringPayload'
          - type: 'null'
          description: 'Result of the structuring stage. Left out when the structuring service hands off
            by reference (`PIPELINE_HANDOFF_BY_REFERENCE`, recommended for KK): this service then reads
            `structuring_results` (and `ocr_results`) of the shared database'
        pipeline_name_sequence:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Pipeline Name Sequence
          description: 'The services this request runs, in order, from the central orchestrator (forwarded
            by the orchestrator and each stage): `guardrails`, `ekstraksi`, `structuring`, `scoring`,
            guardrails optional at the front and the end cut off, never one skipped in the middle. The
            last one ends the request: its result is the answer, as it is, and nothing is handed on. Omitted:
            the full pipeline'
          examples:
          - - guardrails
            - ekstraksi
            - structuring
            - scoring
      type: object
      required:
      - request_id
      title: ScoringJobRequest
      description: Body of `POST /v1/scoring/jobs` (§8.1).
...
    ScoringJobStatusResponse:
      properties:
        status_code:
          type: integer
          title: Status Code
          description: Same as the HTTP status code
          default: 200
          examples:
          - 200
        status_desc:
          type: string
          title: Status Desc
          description: Reason phrase of `status_code`
          default: OK
          examples:
          - OK
        message:
          type: string
          title: Message
          description: Human-readable outcome
          default: Success
          examples:
          - Success
        errors:
          type: 'null'
          title: Errors
          description: Always null on success
        request_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Request Id
          description: 'The request_id this response belongs to: the one sent in the form / body / path,
            otherwise the `X-Request-ID` header, otherwise one generated by the service'
          examples:
          - REQ_9cb01af2-493d-446d-b191-af120333f6d0
        data:
          $ref: '#/components/schemas/ScoringJobStatus'
      type: object
      required:
      - data
      title: ScoringJobStatusResponse
...
    StructuringPayload:
      properties:
        nomor_kk:
          $ref: '#/components/schemas/StructuredFieldPayload'
          description: 'The KK number. Note `nomor_kk`, not `no_kk`: the rename happens in the orchestrator'
        nama_kepala_keluarga:
          $ref: '#/components/schemas/StructuredFieldPayload'
        alamat:
          $ref: '#/components/schemas/StructuredFieldPayload'
        desa_kelurahan:
          $ref: '#/components/schemas/StructuredFieldPayload'
        rt:
          $ref: '#/components/schemas/StructuredFieldPayload'
          description: 'Shares one `ocr_conf` with `rw`: both are cut from the same OCR box'
        rw:
          $ref: '#/components/schemas/StructuredFieldPayload'
        kecamatan:
          $ref: '#/components/schemas/StructuredFieldPayload'
        kabupaten_kota:
          $ref: '#/components/schemas/StructuredFieldPayload'
        provinsi:
          $ref: '#/components/schemas/StructuredFieldPayload'
        kode_pos:
          $ref: '#/components/schemas/StructuredFieldPayload'
        tanggal_dikeluarkan:
          $ref: '#/components/schemas/StructuredFieldPayload'
        anggota_keluarga:
          items:
            $ref: '#/components/schemas/StructuredMemberPayload'
          type: array
          title: Anggota Keluarga
          description: One entry per row on the card, in card order. May be empty, which the validity
            gate rejects
        reject_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Reject Reason
          description: The first rejecting rule of the KK validity gate, in Indonesian; null when accepted.
            It lives in this payload rather than only in the outcome row because the orchestrator is stateless
            and reads stages through their API -- a reason kept elsewhere could never become the client's
            400
          examples:
          - null
      additionalProperties: true
      type: object
      required:
      - nomor_kk
      - nama_kepala_keluarga
      - alamat
      - desa_kelurahan
      - rt
      - rw
      - kecamatan
      - kabupaten_kota
      - provinsi
      - kode_pos
      - tanggal_dikeluarkan
      title: StructuringPayload
      description: Result of the structuring stage (`structuring_results`), forwarded to scoring.
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      description: API key for this service
      in: header
      name: X-API-Key
...